Refactoring: Moved Assembly to its own file
