+++ iOS/Toast Delivery/Environment/Assembly.swift
@@ -1,0 +1,15 @@
+//
+//  Assembly.swift
+//  Toast Delivery
+//
+
+import DITranquillity
+
+protocol Assembly {
+    var container: DIContainer { get }
+}
+extension Assembly {
+    func resolve<T>() -> T {
+        container.resolve()
+    }
+}
